Issue #3009394 by drupal.dinesh, balazswmann: SwaggerUILinkFormatter
