+++ js/swagger-ui-formatter.js
